Update: Split prepareWrite into stream module (closes #193)
