+++ lib/dest/index.js
@@ -1,59 +1,26 @@
 'use strict';
 
-var through2 = require('through2');
-var sourcemaps = require('gulp-sourcemaps');
-var duplexify = require('duplexify');
-var valueOrFunction = require('value-or-function');
+var pumpify = require('pumpify');
+var prepare = require('vinyl-prepare');
 
 var sink = require('../sink');
-var prepareWrite = require('../prepare-write');
-var writeContents = require('./write-contents');
+var saveFile = require('./save-file');
 
 function dest(outFolder, opt) {
   if (!opt) {
     opt = {};
   }
 
-  var sourcemapsOpt = valueOrFunction(
-      ['boolean', 'string', 'object'], opt.sourcemaps);
-
-  function saveFile(file, enc, callback) {
-    prepareWrite(outFolder, file, opt, onPrepare);
-
-    function onPrepare(prepareErr) {
-      if (prepareErr) {
-        return callback(prepareErr);
-      }
-      writeContents(file, callback);
-    }
-  }
-
-  var saveStream = through2.obj(opt, saveFile);
-  if (!sourcemapsOpt) {
-    // Sink the save stream to start flowing
-    // Do this on nextTick, it will flow at slowest speed of piped streams
-    process.nextTick(sink(saveStream));
-
-    return saveStream;
-  }
-
-  if (typeof sourcemapsOpt === 'boolean') {
-    sourcemapsOpt = {};
-  } else if (typeof sourcemapsOpt === 'string') {
-    sourcemapsOpt = {
-      path: sourcemapsOpt,
-    };
-  }
-
-  var mapStream = sourcemaps.write(sourcemapsOpt.path, sourcemapsOpt);
-  var outputStream = duplexify.obj(mapStream, saveStream);
-  mapStream.pipe(saveStream);
+  var saveStream = pumpify.obj(
+    prepare.dest(outFolder, opt),
+    saveFile(opt)
+  );
 
   // Sink the output stream to start flowing
   // Do this on nextTick, it will flow at slowest speed of piped streams
-  process.nextTick(sink(outputStream));
+  process.nextTick(sink(saveStream));
 
-  return outputStream;
+  return saveStream;
 }
 
 module.exports = dest;
