+++ lib/symlink/index.js
@@ -3,13 +3,16 @@
 var path = require('path');
 
 var fs = require('graceful-fs');
-var through2 = require('through2');
-var valueOrFunction = require('value-or-function');
+var pumpify = require('pumpify');
+var through = require('through2');
+var prepare = require('vinyl-prepare');
 var koalas = require('koalas');
+var valueOrFunction = require('value-or-function');
 
+var fo = require('../file-operations');
 var sink = require('../sink');
-var prepareWrite = require('../prepare-write');
 
+var number = valueOrFunction.number;
 var boolean = valueOrFunction.boolean;
 
 function symlink(outFolder, opt) {
@@ -18,20 +21,25 @@ function symlink(outFolder, opt) {
   }
 
   function linkFile(file, enc, callback) {
-    var srcPath = file.path;
+    // Fetch the path as it was before prepare.write()
+    var srcPath = file.history[file.history.length - 2];
     var symType = (file.isDirectory() ? 'dir' : 'file');
     var isRelative = koalas(boolean(opt.relative, file), false);
 
-    prepareWrite(outFolder, file, opt, onPrepare);
+    // This is done inside prepareWrite to use the adjusted file.base property
+    if (isRelative) {
+      srcPath = path.relative(file.base, srcPath);
+    }
 
-    function onPrepare(prepareErr) {
-      if (prepareErr) {
-        return callback(prepareErr);
-      }
+    // TODO: make DRY with .dest()
+    var dirMode = number(opt.dirMode, file);
+    var writeFolder = path.dirname(file.path);
 
-      // This is done inside prepareWrite to use the adjusted file.base property
-      if (isRelative) {
-        srcPath = path.relative(file.base, srcPath);
+    fo.mkdirp(writeFolder, dirMode, onMkdirp);
+
+    function onMkdirp(mkdirpErr) {
+      if (mkdirpErr) {
+        return callback(mkdirpErr);
       }
 
       fs.symlink(srcPath, file.path, symType, onSymlink);
@@ -45,7 +53,10 @@ function symlink(outFolder, opt) {
     }
   }
 
-  var stream = through2.obj(opt, linkFile);
+  var stream = pumpify.obj(
+    prepare.dest(outFolder, opt),
+    through.obj(opt, linkFile)
+  );
   // Sink the stream to start flowing
   // Do this on nextTick, it will flow at slowest speed of piped streams
   process.nextTick(sink(stream));
