+++ package.json
@@ -36,13 +36,15 @@
     "lazystream": "^1.0.0",
     "merge-stream": "^1.0.0",
     "object-assign": "^4.0.0",
+    "pumpify": "^1.3.5",
     "strip-bom": "^2.0.0",
     "strip-bom-stream": "^1.0.0",
     "through2": "^2.0.0",
     "through2-filter": "^2.0.0",
     "vali-date": "^1.0.0",
     "value-or-function": "^1.2.0",
-    "vinyl": "^2.0.0"
+    "vinyl": "^2.0.0",
+    "vinyl-prepare": "gulpjs/vinyl-prepare"
   },
   "devDependencies": {
     "del": "^2.2.0",
