+++ test/dest.js
@@ -108,7 +108,9 @@ describe('.dest()', function() {
     ], done);
   });
 
-  it('accepts the sourcemap option as an object', function(done) {
+  // TODO: actually test sourcemaps
+
+  it.skip('accepts the sourcemap option as an object', function(done) {
     var options = {
       sourcemaps: {
         addComment: false,
